requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
greenlet==1.1.3
idna==3.4
minimalmodbus==2.0.1
paho-mqtt==1.6.1
pyserial==3.5
requests==2.28.1
SQLAlchemy==1.4.41
urllib3==1.26.12
